create function that calculates the total, called within the useEffect hook in the confirmation page
